update card tests and names
