--- @ndlib/gatsby-theme-marble/src/components/Shared/CardGroup/sx.js
@@ -1,11 +1,0 @@
-module.exports = {
-  wrapper: {
-    margin: '0 -1rem',
-  },
-  card: {
-    display: 'inline-block',
-    padding: '1rem',
-    verticalAlign: 'top',
-    width: ['100%', '50%', '33.33%'],
-  },
-}
